pyproject.toml
[build-system]
requires = ["hatchling>=1.18.0"]
build-backend = "hatchling.build"

[project]
name = "drift-bench"
version = "0.1.0"
description = "Benchmark for trajectory drift in multi-turn LLM-assisted scientific ideation"
requires-python = ">=3.11"
dependencies = [
    "litellm>=1.40.0",
    "tenacity>=8.2.0",
    "pyyaml>=6.0",
    "jsonlines>=4.0.0",
    "pandas>=2.0.0",
    "scipy>=1.11.0",
    "matplotlib>=3.8.0",
    "seaborn>=0.13.0",
    "jinja2>=3.1.0",
    "jsonschema>=4.20.0",
    "pyarrow>=14.0.0",
    "scikit-learn>=1.3.0",
    "statsmodels>=0.14.0",
    "krippendorff>=0.6.0",
]

[project.optional-dependencies]
dev = ["pytest>=7.0", "ruff>=0.1.0"]

[project.scripts]
drift-bench = "drift_bench.pipeline:main"
